feat: re-implement oauth
The initial oauth implementation was terrible because yup_oauth2 did not
allow us to opt out of the user interaction flow. This meant that
whenever anything went wrong, our program would hang waiting for a user
to open a browser. This was unacceptable for an mcp server or scripting
tool. So yup_oauth2 had to be removed and google_sheets4 as well due to
coupling.

These were replaced by oauth2 and sheets. This means we wrote a lot of
the oauth workflow code ourselves. Ultimately this gave us what we want:
the only time the program will hang is when we intend it to do so with
tiller auth. tiller auth --verify will not hang, nor will any other
commands.

The auth workflow is verified as working by local testing with my Google
account.